+++ groups/templates/join_group.html
@@ -1,0 +1,6 @@
+<h1>Join Group</h1>
+<form method="post">
+    {% csrf_token %}
+    {{ form.as_p }}
+    <button type="submit">Submit</button>
+</form>
