--- profiles/templates/group_view.html
@@ -1,11 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>{{ group.name }}</h2>
-<h3>Members</h3>
-<ul>
-    {% for member in members %}
-    <li>{{ member.name }}</li>
-    {% endfor %}
-</ul>
-{% endblock %}
