+++ profiles/views.py
@@ -235,160 +235,3 @@ class DeleteRunnerDayData(DeleteView, DataMixin):
         get_best_five_summ.delay()
         return redirect(success_url, success_msg)
 
-
-class AddFamily(CreateView, LoginRequiredMixin):
-    model = Group
-    template_name = 'addfamily.html'
-    form_class = AddFamilyForm
-    success_url = reverse_lazy('profile:profile')
-
-    def get_context_data(self, *args, object_list=None, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        return context
-
-
-
-
-def add_family(request, username):
-    families = Group.objects.all()
-
-    if request.method == 'POST' and Group.objects.filter(runner__runner_status=1):
-        form = FamilyForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('profile:family_list', request.user )
-
-
-
-
-
-    else:
-        form = FamilyForm()
-    return render(request, 'add_family.html', {'form': form, 'families':families})
-
-def family_list(request, username):
-    families = Group.objects.all()
-
-    return render(request, 'family_list.html', {'families': families, 'username':request.user.username})
-
-
-def show_reset(request):
-    form = ResetForm()
-    if request.method == "POST":
-        try:
-            username_form = request.POST.get("username")
-            password = request.POST.get("password1")
-            password2 = request.POST.get("password2")
-            key = request.POST.get("keyword")
-
-            getTeam = Teams.objects.get(team=username_form[:3])
-
-            keywordOfTeam = getTeam.keyword
-
-            try:
-                username = User.objects.get(username=username_form)
-            except ObjectDoesNotExist:
-                messages.error(request,
-                               'Участник с таким номером не найден')
-                render(request, 'pass_reset.html', {'form': form})
-
-            if key != keywordOfTeam:
-                messages.error(request,
-                               'Неверно указано кодовое слово')
-                render(request, 'pass_reset.html', {'form': form})
-
-            if password != password2:
-                messages.error(request,
-                               'Введенные пароли не совпадают')
-                render(request, 'pass_reset.html', {'form': form})
-
-            if key.lower() == keywordOfTeam and User.objects.get(username=username_form) and password == password2:
-                user = User.objects.get(username=username_form)
-                print(user)
-
-                user.set_password(password)
-                user.save()
-                return redirect('authorize:login')
-
-        except:
-            pass
-
-    return render(request, 'pass_reset.html', {'form': form})
-
-
-def show_reset_success(request):
-    return render(request, 'pass_updated.html')
-
-
-
-
-class MyGroup(ListView, DataMixin):
-    model = Group
-    template_name = 'mygroup.html'
-    context_object_name = 'data'
-
-    def get_context_data(self, *args, object_list=None, **kwargs):
-        context = super().get_context_data(**kwargs)
-        group_users = {}
-        # поулчаем группу через юезра из кварг
-        try:
-            obj = User.objects.get(username=self.kwargs['username'])
-
-            group = obj.runner_group
-            if group is not None:
-            # получаем всех пользователей с этой группой
-                group_stat = User.objects.filter(runner_group=obj.runner_group)
-
-
-                group_users[obj.runner_group] = []
-
-                for user in group_stat:
-                    stats_obj = Statistic.objects.get(runner_stat_id=obj.id)
-                    group_users[obj.runner_group].append({
-                        'group': str(group),
-                        'user': user.username,
-                        'total_distance': stats_obj.total_distance,
-                        'total_time': stats_obj.total_time,
-                        'total_average_temp': stats_obj.total_average_temp,
-                        'total_days': stats_obj.total_days,
-                        'total_runs': stats_obj.total_runs,
-                        'total_balls': stats_obj.total_balls,
-                        'is_qualificated': stats_obj.is_qualificated
-                    })
-                context['qs'] = group_users
-            else:
-                context['message']='Вы не состоите в группе'
-
-
-
-        except:
-            context['qs'] = None
-
-        return context
-#добавляем группу участнику если он в ней состоит
-def addRunnerToGroup(request, username):
-    if request.method == 'POST':
-        group_id = request.POST.get('group_id')
-        group = get_object_or_404(User, id=request.user.id).runner_group
-
-        if group:
-            messages.warning(request, f'Вы уже входите  в эту группу: {group}.')
-        else:
-            group.runner_group=group_id
-            messages.success(request, f'Вы создали и добавились в группу{group.group_title}.')
-        return redirect('group_view', group_id=group.id)
-
-    groups = Group.objects.all()
-    return render(request, 'add_runner_to_group.html', {'groups': groups})
-
-
-
-
-#отображение участников группы в профиле
-
-# def my_group(request, username):
-#
-#     print(data)
-#
-#     return render(request, 'mygroup.html',data)
