+++ requirements.txt
@@ -20,6 +20,7 @@ click-repl==0.3.0
 colorama==0.4.6
 config==0.5.1
 cookiecutter==2.6.0
+crispy-bootstrap4==2024.1
 crispy-tailwind==1.0.3
 cron-descriptor==1.4.3
 cryptography==42.0.7
@@ -39,6 +40,7 @@ django-crispy-forms==2.1
 django-debug-toolbar==4.4.2
 django-dramatiq==0.11.6
 django-filter==24.2
+django-hashedfilenamestorage==2.4.1
 django-multiupload==0.6.1
 django-redis==5.2.0
 django-tailwind==3.8.0
