+++ templates/components/header.html
@@ -25,7 +25,7 @@
 
 
                             <a class="dropdown-item text-center"
-                               href="{% url 'profile:mygroup' username=user.username %}">Группа</a>
+                               href="{% url 'groups:mygroup' username=user.username %}">Группа</a>
                             <div class="dropdown-divider"></div>
